+++ ListableUI/Sources/Layout/Flow/FlowListLayout.swift
@@ -601,13 +601,7 @@ final class FlowListLayout : ListLayout {
     {
         // 1) Calculate the base values used to drive the layout.
         
-        let boundsContext = ListContentBounds.Context(
-            viewSize: context.viewBounds.size,
-            safeAreaInsets: context.safeAreaInsets,
-            direction: self.direction
-        )
-        
-        let bounds = self.layoutAppearance.bounds ?? context.environment.listContentBounds(in: boundsContext)
+        let bounds = self.resolvedBounds(in: context)
         
         let spacings = self.layoutAppearance.spacings